12/26/2025: Upgraded some of our user service variables. Now I think I'm getting to a point where I can easily delete

+++ src/main/java/com/monday/monday_backend/auth/users/UserPreferencesEntity.java
@@ -46,11 +46,10 @@ public class UserPreferencesEntity {
 
     @Setter
     @Column(name = "max_chunks_per_session")
-    private Integer maxChunksPerSession;
+    private Long maxChunksPerSession;
 
     @Setter
     @Column(name = "max_tokens_per_session")
-    private Integer maxTokensPerSession;
+    private Long maxTokensPerSession;
 
-    // TODO: Incorporate recall options later.
 }
